# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CoinSphere API
  version: 1.0.0
  description: >-
    The versioned HTTP and WebSocket contract for the CoinSphere modular
    monolith. This document is the source for the generated Go and TypeScript
    contract types.
servers:
  - url: /
    description: Same-origin deployment
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/config/ai-models/{configId}/agent-bindings:
    parameters:
      - $ref: '#/components/parameters/ConfigId'
    put:
      tags: [config]
      operationId: bindAiModelAgents
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/config/assistant-agents:
    get:
      tags: [config]
      operationId: listAssistantAgents
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [config]
      operationId: createAssistantAgent
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/config/assistant-agents/meta:
    get:
      tags: [config]
      operationId: getAssistantAgentMeta
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/config/assistant-agents/{agentId}:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    put:
      tags: [config]
      operationId: updateAssistantAgent
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    patch:
      tags: [config]
      operationId: patchAssistantAgent
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [config]
      operationId: deleteAssistantAgent
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ConfigId: { name: configId, in: path, required: true, schema: { type: integer, format: int64 } }
    AgentId: { name: agentId, in: path, required: true, schema: { type: integer, format: int64 } }
...
  requestBodies:
    JsonObject:
      required: false
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JsonObject'
  responses:
    Success:
      description: Successful data envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
...
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing, expired, revoked, or invalid authentication
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated principal lacks permission or ownership
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
    SuccessResponse:
      type: object
      required: [code, msg, data]
      properties:
        code: { type: integer, format: int32, example: 200 }
        msg: { type: string, example: success }
        data: {}
...
    Problem:
      type: object
      required: [type, title, status, detail, requestId]
      properties:
        type: { type: string, format: uri-reference, example: about:blank }
        title: { type: string, example: Unauthorized }
        status: { type: integer, format: int32, example: 401 }
        detail: { type: string }
        requestId: { type: string }
...